Should apply most Clang-Tidy suggestions

+++ sources/code.cpp
@@ -3,4 +3,4 @@
 /**
  * Returns whether or not the Code is a termination code.
  */
-int is_termination_code(const Code code) { return code == CODE_QUIT || code == CODE_CLOSE; }
+int is_termination_code(const Code code) { return static_cast<int>(code == CODE_QUIT || code == CODE_CLOSE); }
